Created the daemon generator to run the processor in background.
We have used the git://github.com/dougal/daemon_generator.git to generate the files.

=== FIXME: stopping the daemon generates ugly logs ===
By now the stopping of the daemon generates some ugly logs on the script log (daemonized_solr_processor.rb.log)
A little googling has given no useful results. It needs further searching.

Anyway the ugly logs does not seem to be a problem to go on production as its not a blocker bug.